+++ applications/calculations/lib/calculations.js
@@ -109,13 +109,6 @@ p.connection = function (connectionId, sessionId, session) {
 };
 
 /**
- * @see Application#connectionTo
- */
-p.connectionTo = function (clusterMemberId, connectionId, sessionId) {
-  // Do nothing, since this isn't a clustered example application.
-};
-
-/**
  * @see Application#disconnection
  *
  * Note that since this application is configured to use no sessions,
@@ -124,20 +117,6 @@ p.connectionTo = function (clusterMemberId, connectionId, sessionId) {
 p.disconnection = function (connectionId, sessionId) {
   // Do nothing except log it.
   this.thywill.log.debug("Calculations: Client disconnected: " + connectionId);
-};
-
-/**
- * @see Application#disconnectionFrom
- */
-p.disconnectionFrom = function (clusterMemberId, connectionId, sessionId) {
-  // Do nothing, since this isn't a clustered example application.
-};
-
-/**
- * @see Application#clusterMemberDown
- */
-p.clusterMemberDown = function (clusterMemberId, connectionData) {
-  // Do nothing, since this isn't a clustered example application.
 };
 
 //-----------------------------------------------------------